+++ modules/generator/config.go
@@ -19,10 +19,11 @@ const (
 
 // Config for a generator.
 type Config struct {
-	Ring      RingConfig      `yaml:"ring"`
-	Processor ProcessorConfig `yaml:"processor"`
-	Registry  registry.Config `yaml:"registry"`
-	Storage   storage.Config  `yaml:"storage"`
+	Ring       RingConfig      `yaml:"ring"`
+	Processor  ProcessorConfig `yaml:"processor"`
+	Registry   registry.Config `yaml:"registry"`
+	Storage    storage.Config  `yaml:"storage"`
+	MaxSpanAge int64           `yaml:"max_span_age_sec"`
 }
 
 // RegisterFlagsAndApplyDefaults registers the flags.
@@ -31,6 +32,8 @@ func (cfg *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet)
 	cfg.Processor.RegisterFlagsAndApplyDefaults(prefix, f)
 	cfg.Registry.RegisterFlagsAndApplyDefaults(prefix, f)
 	cfg.Storage.RegisterFlagsAndApplyDefaults(prefix, f)
+	// setting default for max span age before discarding to 30 sec
+	cfg.MaxSpanAge = 30
 }
 
 type ProcessorConfig struct {